refactor(commands): extract command wiring into commands/ module with registerAll
Candidate 3 from architecture review: index.js was 548 lines wiring 12+
commands inline with duplicated patterns (resolveLocation → fetchWithCache →
display). Extracted all command registration into src/commands/index.js with
a single registerAll(program) function. index.js is now 15 lines.

Shared fetch-display-cache logic deepened into runCurrentCommand helper,
eliminating 8x duplicated wiring across now/forecast/5day/coords/radar.
Also added --json output support consistently across commands that were
missing it.

Updated watch.test.js to read from new file location.

All 629 tests pass. Lint clean.
